+++ testdata/singleton_track/sclass_of_const_arity_accuses.rb
@@ -1,0 +1,11 @@
+# Same, wrong arity: MRI raises ArgumentError, itaruby reports E0102.
+class Meter
+end
+
+class << Meter
+  def calibrate(scale)
+    scale
+  end
+end
+
+Meter.calibrate(1, 2)